+++ scripts/rollback/rollback-2.10.14.sh
@@ -1,0 +1,166 @@
+#!/bin/bash
+# timestamp: 1620722566165
+
+npm dist-tag add @midwayjs/egg-layer@2.10.13 latest
+npm dist-tag add @midwayjs/egg-layer@2.10.14 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.10.13 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.10.14 latest
+npm dist-tag add @midwayjs/express-layer@2.10.13 latest
+npm dist-tag add @midwayjs/express-layer@2.10.14 beta
+tnpm dist-tag add @midwayjs/express-layer@2.10.13 latest
+tnpm dist-tag add @midwayjs/express-layer@2.10.14 latest
+npm dist-tag add @midwayjs/koa-layer@2.10.13 latest
+npm dist-tag add @midwayjs/koa-layer@2.10.14 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.10.13 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.10.14 latest
+npm dist-tag add @midwayjs/runtime-engine@2.10.12 latest
+npm dist-tag add @midwayjs/runtime-engine@2.10.14 beta
+tnpm dist-tag add @midwayjs/runtime-engine@2.10.12 latest
+tnpm dist-tag add @midwayjs/runtime-engine@2.10.14 latest
+npm dist-tag add @midwayjs/runtime-mock@2.10.12 latest
+npm dist-tag add @midwayjs/runtime-mock@2.10.14 beta
+tnpm dist-tag add @midwayjs/runtime-mock@2.10.12 latest
+tnpm dist-tag add @midwayjs/runtime-mock@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-app@2.10.13 latest
+npm dist-tag add @midwayjs/serverless-app@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.10.13 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@2.10.12 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-aws-starter@2.10.12 latest
+tnpm dist-tag add @midwayjs/serverless-aws-starter@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.10.13 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.10.13 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.10.13 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.10.13 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@2.10.13 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-scf-starter@2.10.13 latest
+tnpm dist-tag add @midwayjs/serverless-scf-starter@2.10.14 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.10.13 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.10.14 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.10.13 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.10.14 latest
+npm dist-tag add @midwayjs/static-layer@2.10.13 latest
+npm dist-tag add @midwayjs/static-layer@2.10.14 beta
+tnpm dist-tag add @midwayjs/static-layer@2.10.13 latest
+tnpm dist-tag add @midwayjs/static-layer@2.10.14 latest
+npm dist-tag add @midwayjs/bootstrap@2.10.13 latest
+npm dist-tag add @midwayjs/bootstrap@2.10.14 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.10.13 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.10.14 latest
+npm dist-tag add @midwayjs/cache@2.10.13 latest
+npm dist-tag add @midwayjs/cache@2.10.14 beta
+tnpm dist-tag add @midwayjs/cache@2.10.13 latest
+tnpm dist-tag add @midwayjs/cache@2.10.14 latest
+npm dist-tag add @midwayjs/consul@2.10.13 latest
+npm dist-tag add @midwayjs/consul@2.10.14 beta
+tnpm dist-tag add @midwayjs/consul@2.10.13 latest
+tnpm dist-tag add @midwayjs/consul@2.10.14 latest
+npm dist-tag add @midwayjs/core@2.10.13 latest
+npm dist-tag add @midwayjs/core@2.10.14 beta
+tnpm dist-tag add @midwayjs/core@2.10.13 latest
+tnpm dist-tag add @midwayjs/core@2.10.14 latest
+npm dist-tag add @midwayjs/decorator@2.10.12 latest
+npm dist-tag add @midwayjs/decorator@2.10.14 beta
+tnpm dist-tag add @midwayjs/decorator@2.10.12 latest
+tnpm dist-tag add @midwayjs/decorator@2.10.14 latest
+npm dist-tag add @midwayjs/faas@2.10.13 latest
+npm dist-tag add @midwayjs/faas@2.10.14 beta
+tnpm dist-tag add @midwayjs/faas@2.10.13 latest
+tnpm dist-tag add @midwayjs/faas@2.10.14 latest
+npm dist-tag add @midwayjs/grpc@2.10.13 latest
+npm dist-tag add @midwayjs/grpc@2.10.14 beta
+tnpm dist-tag add @midwayjs/grpc@2.10.13 latest
+tnpm dist-tag add @midwayjs/grpc@2.10.14 latest
+npm dist-tag add @midwayjs/logger@2.10.13 latest
+npm dist-tag add @midwayjs/logger@2.10.14 beta
+tnpm dist-tag add @midwayjs/logger@2.10.13 latest
+tnpm dist-tag add @midwayjs/logger@2.10.14 latest
+npm dist-tag add midway-schedule@2.10.13 latest
+npm dist-tag add midway-schedule@2.10.14 beta
+tnpm dist-tag add midway-schedule@2.10.13 latest
+tnpm dist-tag add midway-schedule@2.10.14 latest
+npm dist-tag add midway@5.3.13 latest
+npm dist-tag add midway@2.10.14 beta
+tnpm dist-tag add midway@5.3.13 latest
+tnpm dist-tag add midway@2.10.14 latest
+npm dist-tag add @midwayjs/mock@2.10.13 latest
+npm dist-tag add @midwayjs/mock@2.10.14 beta
+tnpm dist-tag add @midwayjs/mock@2.10.13 latest
+tnpm dist-tag add @midwayjs/mock@2.10.14 latest
+npm dist-tag add @midwayjs/prometheus@2.10.13 latest
+npm dist-tag add @midwayjs/prometheus@2.10.14 beta
+tnpm dist-tag add @midwayjs/prometheus@2.10.13 latest
+tnpm dist-tag add @midwayjs/prometheus@2.10.14 latest
+npm dist-tag add @midwayjs/rabbitmq@2.10.13 latest
+npm dist-tag add @midwayjs/rabbitmq@2.10.14 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.10.13 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.10.14 latest
+npm dist-tag add @midwayjs/socketio@2.10.13 latest
+npm dist-tag add @midwayjs/socketio@2.10.14 beta
+tnpm dist-tag add @midwayjs/socketio@2.10.13 latest
+tnpm dist-tag add @midwayjs/socketio@2.10.14 latest
+npm dist-tag add @midwayjs/task@2.10.13 latest
+npm dist-tag add @midwayjs/task@2.10.14 beta
+tnpm dist-tag add @midwayjs/task@2.10.13 latest
+tnpm dist-tag add @midwayjs/task@2.10.14 latest
+npm dist-tag add @midwayjs/typegoose@2.10.13 latest
+npm dist-tag add @midwayjs/typegoose@2.10.14 beta
+tnpm dist-tag add @midwayjs/typegoose@2.10.13 latest
+tnpm dist-tag add @midwayjs/typegoose@2.10.14 latest
+npm dist-tag add @midwayjs/version@2.10.13 latest
+npm dist-tag add @midwayjs/version@2.10.14 beta
+tnpm dist-tag add @midwayjs/version@2.10.13 latest
+tnpm dist-tag add @midwayjs/version@2.10.14 latest
+npm dist-tag add @midwayjs/express@2.10.13 latest
+npm dist-tag add @midwayjs/express@2.10.14 beta
+tnpm dist-tag add @midwayjs/express@2.10.13 latest
+tnpm dist-tag add @midwayjs/express@2.10.14 latest
+npm dist-tag add @midwayjs/koa@2.10.13 latest
+npm dist-tag add @midwayjs/koa@2.10.14 beta
+tnpm dist-tag add @midwayjs/koa@2.10.13 latest
+tnpm dist-tag add @midwayjs/koa@2.10.14 latest
+npm dist-tag add @midwayjs/web@2.10.13 latest
+npm dist-tag add @midwayjs/web@2.10.14 beta
+tnpm dist-tag add @midwayjs/web@2.10.13 latest
+tnpm dist-tag add @midwayjs/web@2.10.14 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.10.13 => 2.10.14
+#  - @midwayjs/express-layer: 2.10.13 => 2.10.14
+#  - @midwayjs/koa-layer: 2.10.13 => 2.10.14
+#  - @midwayjs/runtime-engine: 2.10.12 => 2.10.14
+#  - @midwayjs/runtime-mock: 2.10.12 => 2.10.14
+#  - @midwayjs/serverless-app: 2.10.13 => 2.10.14
+#  - @midwayjs/serverless-aws-starter: 2.10.12 => 2.10.14
+#  - @midwayjs/serverless-fc-starter: 2.10.13 => 2.10.14
+#  - @midwayjs/serverless-fc-trigger: 2.10.13 => 2.10.14
+#  - @midwayjs/serverless-scf-starter: 2.10.13 => 2.10.14
+#  - @midwayjs/serverless-scf-trigger: 2.10.13 => 2.10.14
+#  - @midwayjs/static-layer: 2.10.13 => 2.10.14
+#  - @midwayjs/bootstrap: 2.10.13 => 2.10.14
+#  - @midwayjs/cache: 2.10.13 => 2.10.14
+#  - @midwayjs/consul: 2.10.13 => 2.10.14
+#  - @midwayjs/core: 2.10.13 => 2.10.14
+#  - @midwayjs/decorator: 2.10.12 => 2.10.14
+#  - @midwayjs/faas: 2.10.13 => 2.10.14
+#  - @midwayjs/grpc: 2.10.13 => 2.10.14
+#  - @midwayjs/logger: 2.10.13 => 2.10.14
+#  - midway-schedule: 2.10.13 => 2.10.14
+#  - midway: 5.3.13 => 2.10.14
+#  - @midwayjs/mock: 2.10.13 => 2.10.14
+#  - @midwayjs/prometheus: 2.10.13 => 2.10.14
+#  - @midwayjs/rabbitmq: 2.10.13 => 2.10.14
+#  - @midwayjs/socketio: 2.10.13 => 2.10.14
+#  - @midwayjs/task: 2.10.13 => 2.10.14
+#  - @midwayjs/typegoose: 2.10.13 => 2.10.14
+#  - @midwayjs/version: 2.10.13 => 2.10.14
+#  - @midwayjs/express: 2.10.13 => 2.10.14
+#  - @midwayjs/koa: 2.10.13 => 2.10.14
+#  - @midwayjs/web: 2.10.13 => 2.10.14